yahoo_auctions_parser_eng/requirements.txt
# Core dependencies
beautifulsoup4>=4.12.0  # HTML parsing
python-dotenv>=1.0.0    # Environment variables
requests>=2.31.0        # HTTP requests
lxml>=5.1.0             # XML processing

# Google Sheets integration
gspread>=5.11.2
oauth2client>=4.1.3

# Translation
googletrans>=4.0.0-rc1

# Development tools
mitmproxy>=10.2.2       # For API analysis 